I think I have the file structure fixed again since my last attempt died in a dumpster fire
