+++ GameWebsite/VueFrontend/src/App.vue
@@ -1,9 +1,7 @@
-<template>
-    <div id="app">
-        <router-view />
-    </div>
+﻿<template>
+	<div id="app">
+		<router-view />
+	</div>
 </template>
 
-<style>
-</style>
-
+<style></style>
