--- GameWebsite/VueFrontend/src/components/auth/register.vue
@@ -1,112 +1,0 @@
-<template>
-  <div class="container">
-    <div class="row">
-      <div class="col-lg-6 offset-lg-3 col-sm-10 offset-sm-1">
-        <form
-          class="text-center border border-primary p-5"
-          style="margin-top:70px;height:auto;padding-top:100px !important;"
-          @submit.prevent="registerUser"
-        >
-          <input
-            type="text"
-            id="name"
-            class="form-control mb-5"
-            placeholder="First Name"
-            v-model="register.firstName"
-            required
-          />
-          <input
-            type="text"
-            id="name"
-            class="form-control mb-5"
-            placeholder="Last Name"
-            v-model="register.lastName"
-            required
-          />
-          <input
-            type="text"
-            id="name"
-            class="form-control mb-5"
-            placeholder="NickName"
-            v-model="register.nickName"
-            required
-          />
-          <input
-            type="date"
-            id="name"
-            class="form-control mb-5"
-            placeholder="Birth Date"
-            v-model="register.birthDate"
-            required
-          />
-          <input
-            type="email"
-            id="email"
-            class="form-control mb-5"
-            placeholder="Email"
-            v-model="register.email"
-            required
-          />
-          <!-- Password -->
-          <input
-            type="password"
-            id="password"
-            class="form-control mb-5"
-            placeholder="Password"
-            v-model="register.password"
-          />
-          <p>
-            Already have an account??<router-link to="/"
-              >click here</router-link
-            >
-            <!-- Sign in button -->
-            <center>
-              <button class="btn btn-primary btn-block w-75 my-4" type="submit">
-                Sign in
-              </button>
-            </center>
-          </p>
-        </form>
-      </div>
-    </div>
-  </div>
-</template>
-<script>
-export default {
-  data() {
-    return {
-      register: {
-        firstName: "",
-        lastName: "",
-        birthDate: "",
-        nickName: "",
-        email: "",
-        password: ""
-      }
-    };
-  },
-  methods: {
-    async registerUser() {
-      try {
-        let response = await this.$http.post("http://localhost:3000/users", this.register);
-        //console.log(response);
-        let token = response.data.token;
-        if (token) {
-          localStorage.setItem("jwt", token);
-          this.$router.push("/");
-          alert("Success", "Registration Was successful", "success");
-        } else {
-          alert("Error", "Something Went Wrong", "error");
-        }
-      } catch (err) {
-        let error = err.response;
-        if (error.status == 409) {
-          alert("Error", error.data.message, "error");
-        } else {
-          alert("Error", error.data.err.message, "error");
-        }
-      }
-    }
-  }
-};
-</script>
