+++ GameWebsite/VueFrontend/src/main.js
@@ -1,5 +1,5 @@
 import Vue from "vue";
-import App from "./App.vue";
+import App from "./app.vue";
 import router from "./router";
 import axios from "axios";
 import "bootstrap/dist/css/bootstrap.css";
