+++ GameWebsite/VueFrontend/src/views/game-lobbies/game-lobbies.js
@@ -1,0 +1,23 @@
+export default {
+    name: 'GameLobbies',
+    data() {
+        return {
+            lobbyList: [
+                {
+                    "gameTitle": "Cryptogram Game",
+                    "occupancy": "1",
+                    "capacity": "2",
+                    "gameID": "00001",
+                    "port": "1111"
+                },
+                {
+                    "gameTitle": "Kintsugi",
+                    "occupancy": "3",
+                    "capacity": "4",
+                    "gameID": "00002",
+                    "port": "2222"
+                }
+            ]
+        }
+    }
+};
