+++ GameWebsite/VueFrontend/src/views/game/game.js
@@ -1,0 +1,6 @@
+export default {
+    name: 'Game',
+    props: {
+        lobby: Object
+    }
+};
