--- GameWebsite/VueFrontend/src/views/login.vue
@@ -1,13 +1,0 @@
-<template>
-  <div class="login">
-    <login />
-  </div>
-</template>
-<script>
-import login from "@/components/auth/login";
-export default {
-  components: {
-    login
-  }
-};
-</script>
