+++ GameWebsite/VueFrontend/src/views/login/login.js
@@ -1,0 +1,7 @@
+import login from "../../components/auth/login/login.vue";
+
+export default {
+    components: {
+        login
+    }
+};
