+++ GameWebsite/VueFrontend/src/views/login/login.vue
@@ -1,0 +1,7 @@
+<template>
+	<div class="login">
+		<login />
+	</div>
+</template>
+
+<script src="./login.js"></script>
