--- GameWebsite/VueFrontend/src/views/register.vue
@@ -1,13 +1,0 @@
-<template>
-  <div class="register">
-    <register />
-  </div>
-</template>
-<script>
-import register from "@/components/auth/register";
-export default {
-  components: {
-    register
-  }
-};
-</script>
