+++ GameWebsite/VueFrontend/src/views/register/register.js
@@ -1,0 +1,7 @@
+import register from "../../components/auth/register/register.vue";
+
+export default {
+    components: {
+        register
+    }
+};
