+++ GameWebsite/VueFrontend/src/views/register/register.vue
@@ -1,0 +1,7 @@
+<template>
+	<div class="register">
+		<register />
+	</div>
+</template>
+
+<script src="./register.js"></script>
